+++ tools/splitobs/splitobs.c
@@ -256,48 +256,6 @@ int main(int argc, char** argv) {
     exit_program(NOT_OK,PROGRAM_NAME,PROBLEM,VERSION);
   }
 
-  /* Selon le code dans RMNLIB pour 'wkoffit.c', le chiffre qui sort de 'c_wkoffit' correspond a ceci: */
-#define WKF_INEXISTANT            -3
-#define WKF_VIDE                  -2
-#define WKF_INCONNU               -1
-#define WKF_RANDOM89               1
-#define WKF_SEQUENTIEL89           2
-#define WKF_SEQUENTIELFORTRAN89    3
-#define WKF_CCRN                   4
-#define WKF_CCRN_RPN               5
-#define WKF_BURP                   6
-#define WKF_GRIB                   7
-#define WKF_BUFR                   8
-#define WKF_BLOK                   9
-#define WKF_FORTRAN               10
-#define WKF_COMPRESS              11
-#define WKF_GIF89                 12
-#define WKF_GIF87                 13
-#define WKF_IRIS                  14
-#define WKF_JPG                   15
-#define WKF_KMW                   16
-#define WKF_PBM                   17
-#define WKF_PCL                   18
-#define WKF_PCX                   19
-#define WKF_PDSVICAR              20
-#define WKF_PM                    21
-#define WKF_PPM                   22
-#define WKF_PS                    23
-#define WKF_KMW_                  24
-#define WKF_RRBX                  25
-#define WKF_SUNRAS                26
-#define WKF_TIFF                  27
-#define WKF_UTAHRLE               28
-#define WKF_XBM                   29
-#define WKF_XWD                   30
-#define WKF_ASCII                 31
-#define WKF_BMP                   32
-#define WKF_RANDOM98              33
-#define WKF_SEQUENTIEL98          34
-#define WKF_NETCDF                35
-/* Cette definition est un ajout */
-#define WKF_SQLite                36
-
   status = access(opt.obsin,F_OK);
   if ( status != 0 ) { /* Le fichier existe deja */
       fprintf(stderr,"The file '%s' does not exist and should.\n",opt.obsin);
@@ -305,28 +263,6 @@ int main(int argc, char** argv) {
   }
 
   filetype = c_wkoffit(opt.obsin,strlen(opt.obsin));
-  /* Si le fichier n'est ni BURP ni ASCII alors on suppose que c'est un fichier SQLite */
-  if (filetype != WKF_BURP) {
-    /* Un fichier SQlite commence avec les caracteres 'SQLite format 3' */
-#define SQLITE_FORMAT_STR_LEN     16 /* Longueur de 'SQLite format 3' +1 (pour le caractere nul a la fin de la string) */
-    FILE* file;
-    char line[SQLITE_FORMAT_STR_LEN];
-    char* fgets_return_value;
-
-    file = (FILE*) fopen(opt.obsin,"r");
-    fgets_return_value = fgets(line,SQLITE_FORMAT_STR_LEN,file);
-    fclose(file);
-    if ( fgets_return_value != (char*) NULL ) {
-      if (strcmp(line,"SQLite format 3") == 0)
-        filetype = WKF_SQLite;
-      else
-        filetype = WKF_ASCII;
-    }
-    else {
-      fprintf(stderr,"Cannot determine the file type since there was an error reading the file '%s'\n",opt.obsin);
-      exit_program(NOT_OK,PROGRAM_NAME,PROBLEM,VERSION);
-    }
-  }
 
   /* Si on n'est pas en mode round-robin, alors on a besoin du fichier 'opt.fstin'. */
   if ( opt.roundrobin == 0 ) {
@@ -489,7 +425,7 @@ int main(int argc, char** argv) {
     }
   } /* Fin du 'if ( opt.roundrobin == 0 )' */
 
-  if ( filetype == WKF_SQLite ) {  /* Alors on traite une base de donnees SQL */
+  if ( filetype == WKF_SQLITE3 ) {  /* Alors on traite une base de donnees SQL */
     char sqlreq_tables_without_split_key[MAXSTR];
     char table_list_with_split_key[MAXSTR];
     char table_list_without_split_key[MAXSTR];
@@ -1026,7 +962,7 @@ int main(int argc, char** argv) {
         } /* Fin du 'for (jlatband=0;jlatband<opt.npey;jlatband++)' */
       } /* Fin du 'for (ilonband=0;ilonband<opt.npex;ilonband++)' */
     } /* Fin du 'else' relie au 'if (opt.roundrobin == 0)' */
-  } /* Fin du  if ( filetype == WKF_SQLite ) */
+  } /* Fin du  if ( filetype == WKF_SQLITE3 ) */
   else if ( filetype == WKF_BURP ) {  /* Alors on traite un fichier BURP */
     int i, i_obs_enrgs, i_enrgs, iout = iun+1, nombre_enregistrements, longueur_max_enregistrement, engrs_resume;
     int *adresses = (int*) NULL, *iouts = (int*) NULL;